refactor: переименовал пакеты в camelCase
