+++ app/src/main/java/ru/practicum/android/diploma/core/di/SimilarVacanciesModule.kt
@@ -3,7 +3,7 @@ package ru.practicum.android.diploma.core.di
 import org.koin.androidx.viewmodel.dsl.viewModelOf
 import org.koin.dsl.bind
 import org.koin.dsl.module
-import ru.practicum.android.diploma.similar_vacancies.ui.viewmodel.SimilarVacanciesViewModel
+import ru.practicum.android.diploma.similarVacancies.ui.viewmodel.SimilarVacanciesViewModel
 
 val similarVacanciesModule = module {
     viewModelOf(::SimilarVacanciesViewModel).bind<SimilarVacanciesViewModel>()
