+++ app/src/main/java/ru/practicum/android/diploma/core/di/VacancyDetailsModule.kt
@@ -3,7 +3,7 @@ package ru.practicum.android.diploma.core.di
 import org.koin.androidx.viewmodel.dsl.viewModelOf
 import org.koin.dsl.bind
 import org.koin.dsl.module
-import ru.practicum.android.diploma.vacancy_details.ui.viewmodel.VacancyDetailsViewModel
+import ru.practicum.android.diploma.vacancyDetails.ui.viewmodel.VacancyDetailsViewModel
 
 val vacancyDetailsModule = module {
