+++ app/src/main/java/ru/practicum/android/diploma/search/ui/SearchFragment.kt
@@ -20,7 +20,7 @@ import ru.practicum.android.diploma.R
 import ru.practicum.android.diploma.core.utils.ErrorType
 import ru.practicum.android.diploma.core.utils.adapter.VacancyAdapter
 import ru.practicum.android.diploma.databinding.FragmentSearchBinding
-import ru.practicum.android.diploma.filter.ui.filtration_settings.fragment.FilteringSettingsFragment
+import ru.practicum.android.diploma.filter.ui.filtrationSettings.fragment.FilteringSettingsFragment
 import ru.practicum.android.diploma.search.domain.models.Vacancy
 import ru.practicum.android.diploma.search.ui.state.SearchScreenState
 import ru.practicum.android.diploma.search.ui.viewmodel.SearchViewModel
