+++ app/src/main/java/ru/practicum/android/diploma/vacancyDetails/ui/state/VacancyDetailsScreenState.kt
@@ -1,0 +1,12 @@
+package ru.practicum.android.diploma.vacancyDetails.ui.state
+
+import ru.practicum.android.diploma.core.utils.ErrorType
+import ru.practicum.android.diploma.search.domain.models.VacancyDetails
+
+sealed interface VacancyDetailsScreenState {
+    data object Loading : VacancyDetailsScreenState
+
+    data class Content(val vacancyDetails: VacancyDetails) : VacancyDetailsScreenState
+
+    data class Error(val type: ErrorType) : VacancyDetailsScreenState
+}
