+++ app/src/main/res/layout/fragment_select_country.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".filter.ui.select_country.fragment.SelectCountryFragment">
+    tools:context=".filter.ui.selectCountry.fragment.SelectCountryFragment">
 
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/rightBorder"
