+++ app/src/main/res/layout/fragment_select_industry.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".filter.ui.select_industry.SelectIndustryFragment">
+    tools:context=".filter.ui.selectIndustry.SelectIndustryFragment">
 
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/rightBorder"
