+++ app/src/main/res/layout/fragment_vacancy_details.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fillViewport="false"
-    tools:context=".vacancy_details.ui.fragment.VacancyDetailsFragment">
+    tools:context=".vacancyDetails.ui.fragment.VacancyDetailsFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/details_data"
