+++ app/src/main/res/layout/fragment_workplace.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".filter.ui.select_workplace.fragment.SelectWorkplaceFragment">
+    tools:context=".filter.ui.selectWorkplace.fragment.SelectWorkplaceFragment">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/workplace_toolbar"
