+++ app/src/main/res/navigation/main_nav_graph.xml
@@ -21,7 +21,7 @@
 
     <fragment
         android:id="@+id/filteringSettingsFragment"
-        android:name="ru.practicum.android.diploma.filter.ui.filtration_settings.fragment.FilteringSettingsFragment"
+        android:name="ru.practicum.android.diploma.filter.ui.filtrationSettings.fragment.FilteringSettingsFragment"
         android:label="fragment_filtering_settings"
         tools:layout="@layout/fragment_filtering_settings">
         <action
@@ -42,7 +42,7 @@
 
     <fragment
         android:id="@+id/selectWorkplaceFragment"
-        android:name="ru.practicum.android.diploma.filter.ui.select_workplace.fragment.SelectWorkplaceFragment"
+        android:name="ru.practicum.android.diploma.filter.ui.selectWorkplace.fragment.SelectWorkplaceFragment"
         android:label="fragment_workplace"
         tools:layout="@layout/fragment_workplace">
         <action
@@ -63,13 +63,13 @@
 
     <fragment
         android:id="@+id/selectCountryFragment"
-        android:name="ru.practicum.android.diploma.filter.ui.select_country.fragment.SelectCountryFragment"
+        android:name="ru.practicum.android.diploma.filter.ui.selectCountry.fragment.SelectCountryFragment"
         android:label="fragment_select_country"
         tools:layout="@layout/fragment_select_country" />
 
     <fragment
         android:id="@+id/selectRegionFragment"
-        android:name="ru.practicum.android.diploma.filter.ui.select_region.fragment.SelectRegionFragment"
+        android:name="ru.practicum.android.diploma.filter.ui.selectRegion.fragment.SelectRegionFragment"
         android:label="fragment_select_region"
         tools:layout="@layout/fragment_select_region">
         <argument
@@ -80,13 +80,13 @@
 
     <fragment
         android:id="@+id/selectIndustryFragment"
-        android:name="ru.practicum.android.diploma.filter.ui.select_industry.SelectIndustryFragment"
+        android:name="ru.practicum.android.diploma.filter.ui.selectIndustry.SelectIndustryFragment"
         android:label="fragment_select_industry"
         tools:layout="@layout/fragment_select_industry" />
 
     <fragment
         android:id="@+id/vacancyDetailsFragment"
-        android:name="ru.practicum.android.diploma.vacancy_details.ui.fragment.VacancyDetailsFragment"
+        android:name="ru.practicum.android.diploma.vacancyDetails.ui.fragment.VacancyDetailsFragment"
         android:label="fragment_vacancy"
         tools:layout="@layout/fragment_vacancy_details">
         <action
@@ -111,7 +111,7 @@
 
     <fragment
         android:id="@+id/similarVacanciesFragment"
-        android:name="ru.practicum.android.diploma.similar_vacancies.ui.fragment.SimilarVacanciesFragment"
+        android:name="ru.practicum.android.diploma.similarVacancies.ui.fragment.SimilarVacanciesFragment"
         android:label="fragment_similar_vacancies"
         tools:layout="@layout/fragment_similar_vacancies" >
         <argument
